+++ NEWS.md
@@ -1,4 +1,4 @@
-# guideR (development version)
+# guideR 0.4.0
 
 **New features**